refactor: refactor file structures
