+++ src/content/components/Tab.tsx
@@ -1,8 +1,6 @@
 import { Tab as HeadlessTab } from "@headlessui/react";
 import React from "react";
 
-const Tab = (props: React.ComponentProps<typeof HeadlessTab>) => {
+export const Tab = (props: React.ComponentProps<typeof HeadlessTab>) => {
   return <HeadlessTab className="cursor-pointer" {...props}></HeadlessTab>;
 };
-
-export default Tab;
