+++ src/content/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./Error";
+export * from "./Loading";
+export * from "./PlayButton";
+export * from "./resizable";
+export * from "./Tab";
