+++ src/content/components/resizable.tsx
